requirements.txt
Flask==1.1.2
Flask-Cors==3.0.10
matplotlib==3.3.4
moviepy==1.0.3
numpy==1.20.1
opencv-python==4.5.1.48
Pillow==8.1.2
librosa>=0.8.0
ipython>=7.22.0
